I had a go at a logo (read extended desc.)
The layered one shows the nucleus better with all of the neutrons and
protons.
The logo one is just all black - I didn't want to decide colours with
out it being passed by everyone.
It took a while to get the nucleus right(ish), the outer circle in the
neucleus could be a bit further out and this is how you work out where
they go:
c = no. of objects in circle
r = radius of circle
n = the object you are trying to work out where it goes on the circle (0
is 12 on a clock face, 1 would be the next one round clockwise and so
on)
x=500+(sin(n*(360/c))*(r/sin(90)))
y=500-(sin(90-(n*(360/c)))*(r/sin(90)))

I think, if you want me to explain that better, talk to me in person or
email. Good luck