Modified the default content of index.html to a really crummy login screen
I made a really awful login screen and the CSS is awful. I'm sorry in
advance. Just a note on the logo - for some reason you cannot just embed
an SVG and set the width and height of the embed tag and for the SVG to
size properly. I resorted to using an SVG tag and then setting the
viewBox property to adjust things. It is a messy way of doing it and I
believe the color of the SVG needs updating at some point (I guess this
is Gareth's job).