Wacky colours and simplified
All you have to now is just defince the colours at the top of the page
and it produces it (this could even be dynamic (hint hint server guys).
Also, this shows the pretty spinning insidethat took so long but looks
quite cool. Also, I have scrapped a lot of unnecessary autogenerated
code here: strokes on a circle and all the stuff that goes with a
stroke.