Initial rewrite of the server architecture in Go
This commit changes some of the NodeJS code over to Go code. There are
new installation instructions in the Readme file.

Currently all Node code that handles requests to the server has been
ported over, as has the database functions for dealing with that. The
API code has not yet been ported over, however API requests will work
and return the empty object with basic detail described in the wiki.

This new fork is currently in a state where all client side content
(HTML, JS, CSS, etc) can be modified.