Improved file serving; resolved error with Firefox not working with CSS
Firefox hadn't been loading the CSS previously and it was literally
driving me insane - I had the correct Content-Type and
Chrome/Safari/Opera all seemed to handle it fine.

The bug was that the Content-Type I was submitting was 'text/css'
however it turns out that for it to work in Firefox you must also
include the character set - so 'text/css; charset=utf8' it required
instead.

I've closed the appropriate issue.