Reformatted CSS, login and sign up forms now work
I've managed to get the signup/login forms working and they will now
redirect you to the dashboard if you are logged in. I've also emptied
out index-new.html of any content so that if you aren't logged in it
will fetch the normal login form from welcomedom.html but if you are it
will fetch dashboarddom.html (a sample I've created based on what was
going to be the original dashboard layout - feel free to add to it).